Implemented: reject/recycle button on open and in-progress orders page (#231)
